Added Comparable<BlockLocation> (compareTo) for BlockLocation to get the distance from a block to the given block. Fendinain now checks for line of sights before targeting sapling for pickup or players for begging.
Fenderium:
Now rests after cutting down a tree before moving on to another (30 seconds per a log chopped on the last tree (This will be better refined later on)).
Now looks for the closest tree to itself to chop down.
Now when looking for trees it makes sure it isn't floating.
Much better searching for blocks connected to a tree. It will now usually try to only chop a single tree down at a time now instead of the whole forest.
Now populates the currentTreeBlocks ArrayList when on startExecuting.
Should now properly find the furthest log from the stump it started from.

Added Textures for: fenderium  (Temp logo), logo, treeClearer (Debug item item). Not sure how I forgot to add them to vcs previously.

Almost done with the Fenderium! Need to add the blocks to the inventory (which shouldn't be hard), figure out what to do with those block in the mean time (I have future plans for them, But that is a while away), do a few more tweaks, and then I should be ready to put out another release.